refactor: lift up error boundary to app-router
